+++ qoqo/src/lib.rs
@@ -33,7 +33,7 @@ mod circuit;
 pub use circuit::{convert_into_circuit, CircuitWrapper, OperationIteratorWrapper};
 
 mod quantum_program;
-pub use quantum_program::{QuantumProgramWrapper, convert_into_quantum_program};
+pub use quantum_program::{convert_into_quantum_program, QuantumProgramWrapper};
 
 /// qoqo version information, used for qoqo import/export checks
 pub const QOQO_VERSION: &str = env!("CARGO_PKG_VERSION");